-[CHORE] Integrate the new RetrofitFactory into the old set of classes until replacement and rename.
